+++ public/images/hopemapper-logo.png
